refactor replay playing and league executable functions
