+++ Rofl.UI.Main/Rofl.UI.Main.csproj
@@ -143,6 +143,7 @@
     </Compile>
     <Compile Include="Utilities\DictionaryExtensions.cs" />
     <Compile Include="Utilities\FileAssociations.cs" />
+    <Compile Include="Utilities\ReplayPlayer.cs" />
     <Compile Include="Utilities\ResourceTools.cs" />
     <Compile Include="Views\ExecutableDetailWindow.xaml.cs">
       <DependentUpon>ExecutableDetailWindow.xaml</DependentUpon>
